+++ Catchment/core/isochrone_creator.py
@@ -2,11 +2,11 @@ import json
 import logging
 import os
 from dataclasses import dataclass
+from itertools import groupby
+from operator import itemgetter
 from typing import Dict, List, Optional
 
 import qgis.processing
-from PyQt5.QtCore import QVariant
-from PyQt5.QtNetwork import QNetworkReply
 from qgis.core import (
     QgsCoordinateReferenceSystem,
     QgsCoordinateTransformContext,
@@ -21,7 +21,10 @@ from qgis.core import (
     QgsTask,
     QgsVectorFileWriter,
     QgsVectorLayer,
+    QgsWkbTypes,
 )
+from qgis.PyQt.QtCore import QVariant
+from qgis.PyQt.QtNetwork import QNetworkReply
 
 from ..definitions.constants import Profile, Unit
 from ..qgis_plugin_tools.tools.exceptions import QgsPluginNetworkException
@@ -39,7 +42,10 @@ class IsochroneOpts:
     url: str = ""
     api_key: str = ""
     layer: Optional[QgsVectorLayer] = None
+    polygon_layer: Optional[QgsVectorLayer] = None
     selected_only: bool = False
+    merge_by_field: Optional[QgsField] = None
+    add_walking_field: Optional[QgsField] = None
     distance: Optional[int] = None
     unit: Optional[Unit] = None
     buckets: int = 1
@@ -58,8 +64,10 @@ class IsochroneOpts:
 class IsochroneCreator(QgsTask):
     def __init__(self, opts: IsochroneOpts) -> None:
         self.opts = opts
+        self.error: Optional[Exception] = None
         self.result_layer: Optional[QgsVectorLayer] = None
-        self.points = []
+        self.points: list[QgsFeature] = []
+        self.limiting_polygons: list[QgsFeature] = []
         # no type checking needed, since we check if options are set
         if self.opts.check_if_opts_set():
             self.base_url = self.opts.url
@@ -84,8 +92,9 @@ class IsochroneCreator(QgsTask):
             else:
                 self.params["time_limit"] = 60 * self.opts.distance  # type: ignore
 
-            # reproject layer if needed
+            # reproject layers if needed
             layer: QgsVectorLayer = self.opts.layer
+            polygon_layer: Optional[QgsVectorLayer] = self.opts.polygon_layer
             wgs84 = QgsCoordinateReferenceSystem("EPSG:4326")
             if layer.crs() != wgs84:
                 selected_ids = [
@@ -103,6 +112,22 @@ class IsochroneCreator(QgsTask):
                     "OUTPUT"
                 ]
                 layer.select(selected_ids)
+            if polygon_layer and polygon_layer.crs() != wgs84:
+                MAIN_LOGGER.info(
+                    (
+                        f"Limit polygon layer in {polygon_layer.crs().authid()},"
+                        f" reprojecting to WGS 84 first."
+                    )
+                )
+                alg_params = {
+                    "INPUT": polygon_layer,
+                    "OUTPUT": QgsProcessing.TEMPORARY_OUTPUT,
+                    "TARGET_CRS": wgs84,
+                }
+                polygon_layer = qgis.processing.run(
+                    "native:reprojectlayer", alg_params
+                )["OUTPUT"]
+
             # QgsVectorLayer from main thread may not be used in other threads?
             # How about the QgsFeatures we list here, seems to work fine?
             self.points = (
@@ -110,12 +135,69 @@ class IsochroneCreator(QgsTask):
                 if self.opts.selected_only
                 else list(layer.getFeatures())
             )
+            if polygon_layer:
+                # Store one boundary polygon per point. Store all original polygon ids.
+                fields = QgsFields(polygon_layer.fields())
+                boundary_fid_field = QgsField(
+                    name="fids", type=QVariant.String, typeName="varchar"
+                )
+                fields.append(boundary_fid_field)
+                for point in self.points:
+                    # - In case a point is located inside multiple polygons, consider
+                    #   all of them, i.e. their intersection.
+                    # - In case a point has no boundary polygon, do not limit it.
+                    boundary_polygon = None
+                    for polygon in polygon_layer.getFeatures():
+                        if point.geometry().intersects(polygon.geometry()):
+                            if not boundary_polygon:
+                                # Here the boundary polygon will get all other fields
+                                # from the *first* polygon. Doesn't matter as long as
+                                # we only save the ids in the end
+                                boundary_polygon = QgsFeature(polygon)
+                                boundary_polygon.setFields(fields)
+                                boundary_polygon["fids"] = str(polygon["fid"])
+                            else:
+                                intersection_parts = (
+                                    boundary_polygon.geometry()
+                                    .intersection(polygon.geometry())
+                                    .asGeometryCollection()
+                                )
+                                intersection_geometry = QgsGeometry.fromMultiPolygonXY(
+                                    [
+                                        geometry.asPolygon()
+                                        for geometry in intersection_parts
+                                        if geometry.wkbType() == QgsWkbTypes.Polygon
+                                    ]
+                                )
+                                boundary_polygon.setGeometry(intersection_geometry)
+                                boundary_polygon["fids"] += f",{polygon['fid']}"
+                    self.limiting_polygons.append(boundary_polygon)
+            else:
+                # no limiting polygons for any of the points
+                self.limiting_polygons = len(self.points) * [None]
+
         profile_string = (
             f" by {self.opts.profile.value}" if self.opts.unit == Unit.MINUTES else ""  # type: ignore  # noqa
         )
         direction_string = "to" if self.params["reverse_flow"] else "from"
         selected_string = "selected " if self.opts.selected_only else ""
-        self.name = f"{self.opts.distance} {self.opts.unit.value} {direction_string} {selected_string}{self.opts.layer.name()}{profile_string}"  # type: ignore  # noqa
+        limited_string = (
+            f" limited by {self.opts.polygon_layer.name()}"
+            if self.opts.polygon_layer
+            else ""
+        )
+        merged_string = (
+            f" combined by {self.opts.merge_by_field.name()}"
+            if self.opts.merge_by_field
+            else ""
+        )
+        walking_string = (
+            "with added walking distance " if self.opts.add_walking_field else ""
+        )
+        self.name = (
+            f"{self.opts.distance} {self.opts.unit.value} {walking_string}{direction_string}"  # type: ignore  # noqa
+            f" {selected_string}{self.opts.layer.name()}{profile_string}{limited_string}{merged_string}"  # type: ignore  # noqa
+        )
 
         super().__init__(description=f"Fetching GraphHopper isochrones: {self.name}")
         self.setProgress(0.0)
@@ -132,11 +214,11 @@ class IsochroneCreator(QgsTask):
         """
         try:
             self.result_layer = self.create_isochrone_layer()
-        except QgsPluginNetworkException as e:
-            TASK_LOGGER.error(
-                f"Network request failed, aborting run. Error: {e.message}"  # noqa
-            )
+        except Exception as e:
+            TASK_LOGGER.error(f"Isochrone task failed, aborting run: {repr(e)}")  # noqa
+            self.error = e
             return False
+
         count = self.result_layer.featureCount()
         TASK_LOGGER.info(f"Total of {count} isochrones generated.")
         TASK_LOGGER.info(
@@ -157,13 +239,19 @@ class IsochroneCreator(QgsTask):
         result is the return value from self.run.
         """
         if not result:
-            if not self.result_layer:
-                MAIN_LOGGER.error(
-                    f"Graphhopper request to {self.base_url} failed",
-                    extra={
-                        "details": "Please check your Graphhopper url and your Internet connection."  # noqa
-                    },
-                )
+            if self.error:
+                if isinstance(self.error, QgsPluginNetworkException):
+                    MAIN_LOGGER.error(
+                        f"Graphhopper request to {self.base_url} failed",
+                        extra={
+                            "details": "Please check your Graphhopper url and your Internet connection."  # noqa
+                        },
+                    )
+                else:
+                    MAIN_LOGGER.error(
+                        "Isochrone task failed and returned exception",
+                        extra={"details": repr(self.error)},
+                    )
             elif len(self.points):
                 MAIN_LOGGER.error(
                     "No results, no roads found close to any of the points",
@@ -183,14 +271,50 @@ class IsochroneCreator(QgsTask):
             root = QgsProject.instance().layerTreeRoot()
             root.insertChildNode(1, QgsLayerTreeLayer(self.result_layer))
 
-    def __fetch_bucketed_isochrones(self, point: QgsFeature) -> List[Dict]:
+    def __add_walking_distance(
+        self, isochrone_params: Dict, walking_distance: int
+    ) -> Dict:
+        # Each point may have fixed internal walking distance in meters that has to
+        # be traversed before reaching the entrance, i.e. the Graphhopper network.
+        # This is taken into account to determine the distance to fetch. Note that
+        # this will result in very ugly bucket divisions, so this is best used without
+        # buckets.
+        if not walking_distance:
+            return isochrone_params
+        if self.opts.unit == Unit.METERS:
+            distance = isochrone_params["distance_limit"] - walking_distance
+            if distance < 0:
+                distance = 0
+            TASK_LOGGER.info(
+                f"Added walking distance {walking_distance} m. Fetching isochrone"
+                f" for distance {distance} m."
+            )
+            return {**isochrone_params, "distance_limit": distance}
+        elif self.opts.unit == Unit.MINUTES:
+            # distance in seconds, walking distance in meters, walking speed 5 km/h
+            time = int(
+                isochrone_params["time_limit"] - walking_distance / (5000 / 3600)
+            )
+            if time < 0:
+                time = 0
+            TASK_LOGGER.info(
+                f"Added walking time corresponding to {walking_distance} m. Fetching"
+                f" isochrone for time {time} s."
+            )
+            return {**isochrone_params, "time_limit": time}
+
+    def __fetch_bucketed_isochrones(self, point_feature: QgsFeature) -> List[Dict]:
         # the API may return multiple isochrones for a single point (buckets)
-        geometry = point.geometry()
+        geometry = point_feature.geometry()
         isochrones = []
         # the geometry may be multipoint, handle each point
         for point in geometry.parts():
             isochrone_params = self.params
             isochrone_params["point"] = f"{point.y()},{point.x()}"
+            if self.opts.add_walking_field:
+                isochrone_params = self.__add_walking_distance(
+                    isochrone_params, point_feature[self.opts.add_walking_field.name()]
+                )
             try:
                 isochrone_json = fetch(self.base_url, params=isochrone_params)
             except QgsPluginNetworkException as e:
@@ -217,23 +341,32 @@ class IsochroneCreator(QgsTask):
 
     def __add_isochrones_to_layer(self, layer: QgsVectorLayer) -> None:
         TASK_LOGGER.info("Starting isochrone fetch...")
-        for idx, point in enumerate(self.points):
+        for idx, (point, boundary) in enumerate(
+            zip(self.points, self.limiting_polygons)
+        ):
             bucketed_isochrones = self.__fetch_bucketed_isochrones(point)
             for polygon_in_bucket in bucketed_isochrones:
                 feature = QgsFeature(layer.fields())
-                # save the original feature id separately
+                # when merging, we have to discard extra attributes
+                if self.opts.merge_by_field:
+                    attributes = [
+                        point.id(),
+                        point.attribute(self.opts.merge_by_field.name()),
+                    ]
+                else:
+                    attributes = point.attributes()
                 # setAttributes cannot be used, will destroy any extra fields!!
-                for index, attribute in enumerate(point.attributes()):
+                for index, attribute in enumerate(attributes):
                     feature.setAttribute(index, attribute)
                 # set the added distance field separately
                 bucket = polygon_in_bucket["properties"]["bucket"]
                 distance = (bucket + 1) * (
                     self.opts.distance / self.opts.buckets  # type: ignore
                 )
-                feature.setAttribute("isochrone_distance", distance)
+                feature["isochrone_distance"] = distance
 
-                feature.setGeometry(
-                    QgsGeometry.fromPolygonXY(
+                isochrone = QgsGeometry.fromMultiPolygonXY(
+                    [
                         [
                             [
                                 QgsPointXY(pt[0], pt[1])
@@ -242,8 +375,30 @@ class IsochroneCreator(QgsTask):
                                 ]
                             ]
                         ]
-                    )
+                    ]
                 )
+                if boundary:
+                    feature["boundary_fids"] = boundary["fids"]
+                    isochrone_parts = (
+                        boundary.geometry()
+                        .intersection(isochrone)
+                        .asGeometryCollection()
+                    )
+                    # After intersecting with the boundary, the isochrone may be a
+                    # GeometryCollection of Polygons, LineStrings and Points. We are
+                    # only interested in 2D areas, so collect all the Polygons to a
+                    # MultiPolygon.
+                    isochrone = QgsGeometry.fromMultiPolygonXY(
+                        [
+                            geometry.asPolygon()
+                            for geometry in isochrone_parts
+                            if geometry.wkbType() == QgsWkbTypes.Polygon
+                        ]
+                    )
+                else:
+                    feature["boundary_fids"] = ""
+
+                feature.setGeometry(isochrone)
                 layer.dataProvider().addFeature(feature)
             if idx and idx % 10 == 0:
                 TASK_LOGGER.info(
@@ -256,6 +411,55 @@ class IsochroneCreator(QgsTask):
                 break
             self.setProgress(100 * (idx / len(self.points)))
 
+    def __merge_isochrones_in_layer(self, layer: QgsVectorLayer) -> None:
+        field_name = self.opts.merge_by_field.name()
+        if field_name == "fid":
+            # group by original fid, just in case it is not unique for some reason
+            field_name = "original_fid"
+        TASK_LOGGER.info(f"Merging isochrones by {field_name} value...")
+        # merge isochrones with same field value *and* same distance
+        merge_criterion = itemgetter(
+            layer.dataProvider().fieldNameIndex(field_name),
+            layer.dataProvider().fieldNameIndex("isochrone_distance"),
+        )
+        sorted_features = sorted(layer.getFeatures(), key=merge_criterion)
+        grouped_features = groupby(sorted_features, key=merge_criterion)
+        merged_features = []
+        for group in grouped_features:
+            merged_feature = QgsFeature(layer.fields())
+            merged_geometry = None
+            merged_ids = set()
+            merged_boundary_ids = set()
+            for feature in group[1]:
+                if not merged_geometry:
+                    merged_geometry = feature.geometry()
+                else:
+                    merged_geometry = merged_geometry.combine(feature.geometry())
+                    # now, the result may be polygon *or* multipolygon
+                    if merged_geometry.wkbType() == QgsWkbTypes.Polygon:
+                        merged_geometry = QgsGeometry.fromMultiPolygonXY(
+                            [merged_geometry.asPolygon()]
+                        )
+                merged_ids.add(feature["original_fid"])
+                # boundary fids may be the same, only save different boundary ids
+                merged_boundary_ids.update(feature["boundary_fids"].split(","))
+
+            field_value = group[0][0]
+            distance_value = group[0][1]
+            # finally, sort the ids
+            merged_ids = sorted(list(merged_ids))
+            merged_boundary_ids = sorted(list(merged_boundary_ids))
+            merged_feature.setAttribute("original_fid", ",".join(merged_ids))
+            merged_feature.setAttribute(field_name, field_value)
+            merged_feature.setAttribute("isochrone_distance", distance_value)
+            merged_feature.setAttribute("boundary_fids", ",".join(merged_boundary_ids))
+            merged_feature.setGeometry(merged_geometry)
+            merged_features.append(merged_feature)
+
+        # empty the layer and add new features
+        layer.dataProvider().truncate()
+        layer.dataProvider().addFeatures(merged_features)
+
     def create_isochrone_layer(self) -> QgsVectorLayer:
         """Creates a polygon QgsVectorLayer containing isochrones for points"""
         isochrone_layer = QgsVectorLayer(
@@ -263,18 +467,43 @@ class IsochroneCreator(QgsTask):
         )
 
         # add all the required fields to the new layer
-        fields = QgsFields(self.opts.layer.fields())  # type: ignore
-        # save original feature id separate from new feature id
-        fields.rename(0, "original_fid")
+        fields = QgsFields()
+        # save original fid(s) as string to support multiple point ids
+        original_fid_field = QgsField(
+            name="original_fid", type=QVariant.String, typeName="varchar"
+        )
+        fields.append(original_fid_field)
+
+        if self.opts.merge_by_field:
+            # If isochrones are merged, they will lose all their attributes
+            # other than the one that is used to merge
+            fields.append(self.opts.merge_by_field)
+        else:
+            # We must make a copy of original fields, then edit it, then iterate it
+            # to get the desired final field ordering. Yeah, fields can only be
+            # added to the end, go figure.
+            original_fields = QgsFields(self.opts.layer.fields())  # type: ignore
+            # remove the fid field
+            original_fields.remove(0)
+            for field in original_fields:
+                fields.append(field)
+
+        # add our extra fields last
         distance_field = QgsField(
             name="isochrone_distance", type=QVariant.Double, typeName="double"
         )
+        boundary_fid_field = QgsField(
+            name="boundary_fids", type=QVariant.String, typeName="varchar"
+        )
         fields.append(distance_field)
+        fields.append(boundary_fid_field)
         provider = isochrone_layer.dataProvider()
         provider.addAttributes(fields)
         isochrone_layer.updateFields()
 
         self.__add_isochrones_to_layer(isochrone_layer)
+        if self.opts.merge_by_field:
+            self.__merge_isochrones_in_layer(isochrone_layer)
         # update layer's extent when new features have been added
         isochrone_layer.updateExtents()