+++ Catchment/test/conftest.py
@@ -3,8 +3,9 @@
 """
 This class contains fixtures and common helper function to keep the test files shorter
 """
+import json
 import os
-from typing import Callable, Dict, Optional
+from typing import Callable, Dict, List, Optional, Union
 
 import pytest
 from PyQt5.QtCore import QVariant
@@ -15,7 +16,9 @@ from qgis.core import (
     QgsField,
     QgsFields,
     QgsGeometry,
+    QgsLineString,
     QgsPointXY,
+    QgsPolygon,
     QgsVectorLayer,
 )
 
@@ -23,44 +26,62 @@ from Catchment.core.isochrone_creator import IsochroneOpts
 from Catchment.definitions.constants import Profile, Unit
 from Catchment.plugin import Plugin
 
-from ..qgis_plugin_tools.testing.utilities import get_qgis_app
 from ..qgis_plugin_tools.tools.exceptions import QgsPluginNetworkException
 from ..qgis_plugin_tools.tools.i18n import tr
 
-QGIS_APP, CANVAS, IFACE, PARENT = get_qgis_app()
 MOCK_URL = "http://mock.url"
-
-
-@pytest.fixture(autouse=True)
-def new_project() -> None:
-    """Initializes the QGIS project by removing layers and relations etc."""  # noqa E501
-    # yields nothing
-    yield IFACE.newProject()
 
 
 @pytest.fixture(scope="function")
 def mock_fetch(mocker, request) -> None:
-    """Makes fetch return JSON for a specified URL, exception otherwise.
-    Use by calling mock_fetch(desired_url, json_file_name, error_desired) in a test.
+    """Makes fetch return JSON(s) (and optional error(s)) for specified URL(s).
+    Use by calling mock_fetch(desired_url(s), json_file_name(s), error_desired, required_params(s)) in a test.
     """
 
     def _mock_fetch(
-        url: str,
-        json_to_return: str = "isochrones.json",
-        error: bool = False,
+        desired_url: Union[str, List[str]],
+        json_to_return: Union[str, List[str]] = "isochrones.json",
+        error: Union[bool, List[bool]] = False,
+        required_params: Union[
+            Optional[Dict[str, str]], List[Optional[Dict[str, str]]]
+        ] = None,
     ) -> Callable:
+        if isinstance(desired_url, str):
+            desired_url = [desired_url]
+        if isinstance(json_to_return, str):
+            json_to_return = len(desired_url) * [json_to_return]
+        if isinstance(error, bool):
+            error = len(desired_url) * [error]
+        if isinstance(required_params, dict) or not required_params:
+            required_params = len(desired_url) * [required_params]
+
         def mocked_fetch(
-            incoming_url: str,
+            url: str,
             params: Optional[Dict[str, str]] = None,
         ) -> str:
-            if incoming_url == url:
-                with open(
-                    os.path.join(request.fspath.dirname, "fixtures", json_to_return)
-                ) as f:
-                    # mock error if desired
-                    if error:
-                        raise QgsPluginNetworkException(f.read(), error=302)
-                    return f.read()
+            indices = [
+                idx
+                for idx, matching_url in enumerate(desired_url)
+                if url == matching_url
+            ]
+            # return first matching url+parameter combination
+            for index in indices:
+                # only check parameters if we require specific params
+                if not required_params[index] or all(
+                    [
+                        params.get(key, None) == required_params[index][key]
+                        for key in required_params[index].keys()
+                    ]
+                ):
+                    with open(
+                        os.path.join(
+                            request.fspath.dirname, "fixtures", json_to_return[index]
+                        )
+                    ) as f:
+                        # mock error if desired
+                        if error[index]:
+                            raise QgsPluginNetworkException(f.read(), error=302)
+                        return f.read()
             raise QgsPluginNetworkException(tr("Request failed"))
 
         mocker.patch("Catchment.core.isochrone_creator.fetch", new=mocked_fetch)
@@ -70,28 +91,123 @@ def mock_fetch(mocker, request) -> None:
 
 @pytest.fixture(scope="function")
 def point() -> None:
-    yield QgsPointXY(1.0, 1.0)
+    yield QgsGeometry.fromPointXY(QgsPointXY(1.0, 1.0))
+
+
+@pytest.fixture(scope="function")
+def another_point() -> None:
+    yield QgsGeometry.fromPointXY(QgsPointXY(0.9, 0.9))
+
+
+@pytest.fixture(scope="function")
+def square() -> None:
+    yield QgsGeometry.fromPolygonXY(
+        [
+            [
+                QgsPointXY(0.0, 0.0),
+                QgsPointXY(2.0, 0.0),
+                QgsPointXY(2.0, 2.0),
+                QgsPointXY(0.0, 2.0),
+            ]
+        ]
+    )
+
+
+@pytest.fixture(scope="function")
+def multipolygon() -> None:
+    yield QgsGeometry.fromMultiPolygonXY(
+        [
+            [
+                [
+                    QgsPointXY(0.0, 0.0),
+                    QgsPointXY(2.0, 0.0),
+                    QgsPointXY(2.0, 2.0),
+                    QgsPointXY(0.0, 2.0),
+                ]
+            ],
+            [
+                [
+                    QgsPointXY(4.0, 4.0),
+                    QgsPointXY(6.0, 4.0),
+                    QgsPointXY(6.0, 6.0),
+                    QgsPointXY(4.0, 6.0),
+                ]
+            ],
+        ]
+    )
+
+
+@pytest.fixture(scope="function")
+def triangle() -> None:
+    yield QgsGeometry.fromPolygonXY(
+        [[QgsPointXY(-1.0, -1.0), QgsPointXY(3.0, -1.0), QgsPointXY(1.0, 2.0)]]
+    )
 
 
 @pytest.fixture(scope="function")
 def fields() -> None:
     fields = QgsFields()
-    fields.append(QgsField("id", QVariant.Int))
+    fields.append(QgsField("fid", QVariant.Int))
     fields.append(QgsField("name", QVariant.String))
+    fields.append(QgsField("extra_info", QVariant.String))
+    fields.append(QgsField("extra_field_1", QVariant.Int))
+    fields.append(QgsField("extra_field_2", QVariant.Int))
     yield fields
 
 
 @pytest.fixture(scope="function")
 def point_feature(fields, point) -> None:
     feature = QgsFeature(fields)
-    feature.setGeometry(QgsGeometry.fromPointXY(point))
-    feature.setAttribute("id", 1)
+    feature.setGeometry(point)
+    feature.setAttribute("fid", 1)
     feature.setAttribute("name", "school")
+    feature.setAttribute("extra_info", "first_feature")
+    feature.setAttribute("extra_field_1", 2)
+    feature.setAttribute("extra_field_2", 3)
     yield feature
 
 
 @pytest.fixture(scope="function")
-def vector_layer(fields, point_feature) -> None:
+def another_point_feature(fields, another_point) -> None:
+    feature = QgsFeature(fields)
+    feature.setGeometry(another_point)
+    feature.setAttribute("fid", 2)
+    feature.setAttribute("name", "school")
+    feature.setAttribute("extra_info", "second_feature")
+    feature.setAttribute("extra_field_1", 2)
+    feature.setAttribute("extra_field_2", 4)
+    yield feature
+
+
+@pytest.fixture(scope="function")
+def square_feature(fields, square) -> None:
+    feature = QgsFeature(fields)
+    feature.setGeometry(square)
+    feature.setAttribute("fid", 1)
+    feature.setAttribute("name", "square_school_area_boundary")
+    yield feature
+
+
+@pytest.fixture(scope="function")
+def multipolygon_feature(fields, multipolygon) -> None:
+    feature = QgsFeature(fields)
+    feature.setGeometry(multipolygon)
+    feature.setAttribute("fid", 1)
+    feature.setAttribute("name", "multipolygon_school_area_boundary")
+    yield feature
+
+
+@pytest.fixture(scope="function")
+def triangle_feature(fields, triangle) -> None:
+    feature = QgsFeature(fields)
+    feature.setGeometry(triangle)
+    feature.setAttribute("fid", 1)
+    feature.setAttribute("name", "triangular_school_area_boundary")
+    yield feature
+
+
+@pytest.fixture(scope="function")
+def point_layer(fields, point_feature) -> None:
     layer = QgsVectorLayer("Point?crs=epsg:4326&index=yes", "test_points", "memory")
     provider = layer.dataProvider()
     provider.addAttributes(fields)
@@ -102,10 +218,75 @@ def vector_layer(fields, point_feature) -> None:
 
 
 @pytest.fixture(scope="function")
-def isochrone_opts(vector_layer) -> None:
+def two_point_layer(fields, point_feature, another_point_feature) -> None:
+    layer = QgsVectorLayer("Point?crs=epsg:4326&index=yes", "test_points", "memory")
+    provider = layer.dataProvider()
+    provider.addAttributes(fields)
+    layer.updateFields()
+    provider.addFeature(point_feature)
+    provider.addFeature(another_point_feature)
+    layer.updateExtents()
+    yield layer
+
+
+@pytest.fixture(scope="function")
+def square_layer(fields, square_feature) -> None:
+    layer = QgsVectorLayer(
+        "Polygon?crs=epsg:4326&index=yes", "test_boundaries", "memory"
+    )
+    provider = layer.dataProvider()
+    provider.addAttributes(fields)
+    layer.updateFields()
+    provider.addFeature(square_feature)
+    layer.updateExtents()
+    yield layer
+
+
+@pytest.fixture(scope="function")
+def multipolygon_layer(fields, multipolygon_feature) -> None:
+    layer = QgsVectorLayer(
+        "Polygon?crs=epsg:4326&index=yes", "test_boundaries", "memory"
+    )
+    provider = layer.dataProvider()
+    provider.addAttributes(fields)
+    layer.updateFields()
+    provider.addFeature(multipolygon_feature)
+    layer.updateExtents()
+    yield layer
+
+
+@pytest.fixture(scope="function")
+def triangle_layer(fields, triangle_feature) -> None:
+    layer = QgsVectorLayer(
+        "Polygon?crs=epsg:4326&index=yes", "test_boundaries", "memory"
+    )
+    provider = layer.dataProvider()
+    provider.addAttributes(fields)
+    layer.updateFields()
+    provider.addFeature(triangle_feature)
+    layer.updateExtents()
+    yield layer
+
+
+@pytest.fixture(scope="function")
+def square_plus_triangle_layer(fields, square_feature, triangle_feature) -> None:
+    layer = QgsVectorLayer(
+        "Polygon?crs=epsg:4326&index=yes", "test_boundaries", "memory"
+    )
+    provider = layer.dataProvider()
+    provider.addAttributes(fields)
+    layer.updateFields()
+    provider.addFeature(square_feature)
+    provider.addFeature(triangle_feature)
+    layer.updateExtents()
+    yield layer
+
+
+@pytest.fixture(scope="function")
+def isochrone_opts(point_layer, request) -> None:
     opts = IsochroneOpts(
         url=MOCK_URL,
-        layer=vector_layer,
+        layer=point_layer,
         distance=30,
         unit=Unit.MINUTES,
         profile=Profile.WALKING,
@@ -114,8 +295,8 @@ def isochrone_opts(vector_layer) -> None:
 
 
 @pytest.fixture(scope="function")
-def new_plugin(isochrone_opts) -> None:
-    plugin = Plugin(IFACE)
+def new_plugin(qgis_iface, isochrone_opts) -> None:
+    plugin = Plugin(qgis_iface)
     plugin.initGui()
     # mock options, since mock QgisInterface does not support QgsMapLayerComboBox
     plugin.dlg.read_isochrone_options = lambda: isochrone_opts